Introducing FanControl
I have very strange problem with NBFC causing very huge DPC latency in
my system. So I decided to get all neccessary things for
NBFC/ec-probe.exe to work then created instead a powershell script to
control the fans based on OpenHardware monitor output. This is very
barebones and since I'm nub with Powershell scripting, I'm not sure
about the efficiency of the script. It doesn't consume much of CPU in
task manager ~1% or less. So I'll stick with it. I might refactor it
once I find a much efficient way. Thanks for the binary anyways for
ec-probe @hirschmann.